Move amiibo data management to its own class, and introduce a static field in Program.cs as a global to be used by AmiiboWindow.cs
Data management occurs in background threads as to prevent slowdown of application launch.

Amiibo data updates upon launch of the program as opposed to launch of the window, and data is displayed much quicker than it was before.
